# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Authenticated AI JobMatch API for Pakistan-based job seekers
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CandidateProfile {
  id: number;
  fullName: string;
  email: string;
  preferredLocation: string;
  country: string;
  createdAt: string;
  updatedAt: string;
}

export interface CandidateProfileInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  fullName: string;
  /** @maxLength 120 */
  email: string;
  /**
     * @minLength 2
     * @maxLength 100
     */
  preferredLocation: string;
  /** @pattern ^[A-Za-z]{2}$ */
  country: string;
}

export interface ResumeProfile {
  id: number;
  professionalSummary: string;
  skills: string[];
  education: string;
  experience: string;
  /** @nullable */
  resumeFileName?: string | null;
  /** @nullable */
  extractedTextPreview?: string | null;
  updatedAt: string;
}

export interface ResumeProfileInput {
  /** @maxLength 5000 */
  professionalSummary: string;
  /**
     * @maxItems 100
     * @items.minLength 1
     * @items.maxLength 80
     */
  skills: string[];
  /** @maxLength 5000 */
  education: string;
  /** @maxLength 10000 */
  experience: string;
}

export type ResumeUploadRequestContentType = typeof ResumeUploadRequestContentType[keyof typeof ResumeUploadRequestContentType];


export const ResumeUploadRequestContentType = {
  'application/pdf': 'application/pdf',
  'application/octet-stream': 'application/octet-stream',
} as const;

export interface ResumeUploadRequest {
  /**
     * @minLength 1
     * @maxLength 255
     */
  fileName: string;
  contentType: ResumeUploadRequestContentType;
  /**
     * @minimum 1
     * @maximum 5242880
     */
  size: number;
}

export interface ResumeUploadTarget {
  uploadUrl: string;
  objectPath: string;
}

export interface ResumeUploadCompleteInput {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  objectPath: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  fileName: string;
}

export interface ResumeUploadResult {
  fileName: string;
  extractedTextPreview: string;
  resume: ResumeProfile;
}

export interface Job {
  id: string;
  title: string;
  company: string;
  description: string;
  location: string;
  /** @nullable */
  country?: string | null;
  jobType: string;
  workMode: string;
  requiredSkills: string[];
  applyUrl: string;
  source: string;
  /** @nullable */
  postedAt?: string | null;
}

export interface LiveJobSearchResult {
  jobs: Job[];
  location: string;
  country: string;
  page: number;
}

export type JobMatch = Job & {
  /**
     * @minimum 0
     * @maximum 100
     */
  matchPercentage: number;
  matchedSkills: string[];
  missingSkills: string[];
};

export interface SkillCount {
  skill: string;
  jobs: number;
}

export interface DashboardSummary {
  /**
     * @minimum 0
     * @maximum 100
     */
  profileCompleteness: number;
  localJobs: number;
  matchedJobs: number;
  topSkills: SkillCount[];
}

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Authentication required
 */
export type UnauthorizedResponse = ErrorResponse;

export type LocationParameter = string;

export type JobTypeParameter = typeof JobTypeParameter[keyof typeof JobTypeParameter];


export const JobTypeParameter = {
  'Full-time': 'Full-time',
  'Part-time': 'Part-time',
  Internship: 'Internship',
  Contract: 'Contract',
} as const;

export type WorkModeParameter = typeof WorkModeParameter[keyof typeof WorkModeParameter];


export const WorkModeParameter = {
  'On-site': 'On-site',
  Hybrid: 'Hybrid',
  Remote: 'Remote',
} as const;

export type ListJobsParams = {
/**
 * @maxLength 100
 */
location?: LocationParameter;
jobType?: JobTypeParameter;
workMode?: WorkModeParameter;
};

export type SearchLiveJobsParams = {
/**
 * @minLength 1
 * @maxLength 100
 */
query?: string;
/**
 * @maxLength 100
 */
location?: LocationParameter;
/**
 * ISO 3166-1 alpha-2 country code
 * @pattern ^[A-Za-z]{2}$
 */
country?: string;
jobType?: JobTypeParameter;
workMode?: WorkModeParameter;
/**
 * @minimum 1
 * @maximum 10
 */
page?: number;
};