# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Authenticated AI JobMatch API for Pakistan-based job seekers
 * OpenAPI spec version: 1.0.0
 */
import type { SkillCount } from './skillCount';

export interface DashboardSummary {
  /**
     * @minimum 0
     * @maximum 100
     */
  profileCompleteness: number;
  localJobs: number;
  matchedJobs: number;
  topSkills: SkillCount[];
}