# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Authenticated AI JobMatch API for Pakistan-based job seekers
 * OpenAPI spec version: 1.0.0
 */

export interface HealthStatus {
  status: string;
}