# lib/api-zod/src/generated/types/resumeProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Authenticated AI JobMatch API for Pakistan-based job seekers
 * OpenAPI spec version: 1.0.0
 */

export interface ResumeProfile {
  id: number;
  professionalSummary: string;
  skills: string[];
  education: string;
  experience: string;
  /** @nullable */
  resumeFileName?: string | null;
  /** @nullable */
  extractedTextPreview?: string | null;
  updatedAt: Date;
}